--- src/types/authCredential.d.ts
@@ -1,4 +1,0 @@
-export type AuthCredential = {
-  email: string;
-  password: string;
-};
